Fix test

+++ serializer/user_test.go
@@ -13,7 +13,7 @@ func TestSerializeBaseUsers(t *testing.T) {
 	user.PasswordHash = "testPass"
 
 	userResource := SerializeBaseUsers(1, []*db.User{user})
-	assert.Equal(t, 1, userResource.TotalCount)
+	assert.Equal(t, int64(1), userResource.TotalCount)
 	assert.Equal(t, int64(1), userResource.Data[0].Id)
 	assert.Equal(t, "test", userResource.Data[0].LoginName)
 }
